Skip bad files entries
If the name is an empty string, discard that files entry instead
of rejecting the entire share target.

If an accept entry is invalid, discard the files entry instead of
rejecting either the single accept entry or the entire share target.

(We could simply discard the single accept entry, except if this
results in an empty accept array, as that would incorrectly imply
that all files are accepted. It is better to avoid special cases.)

resolves #77