Associate file extensions and MIME types
An implementation may associate a file extension with
a MIME type. In this case, it is sufficient for a manifest
to only contain the file extension, even if the implementation
only supports filtering on MIME types.

Similarly, an implementation may associate a MIME type with
one or more file extensions. In this case, it is sufficient
for a manifest to only contain a MIME type (optionally with
wildcard subtype), even if the implementation only supports
filtering of file extensions.